Devex 864: split unit tests - formattedTrialSessions
